Global set cli verbosity, escape handling and added security (#14)
Improved security by supporting SecureString as input for Azure CLI parameter. The SecureString value will be passed
on as plain text to Azure CLI, but will be printed as ******** on the Invoke-AzCli verbose output.
Added -EscapeHandling parameter to set automatic escaping of strings on the command-line to Azure CLI. Set to Always
to escape \ and " with the \ escape character. Do not set it, or set it to None, to not escape. This is the default
behavior.
Allow to set global CliVerbosity preference through the $AzCliVerbosityPreference variable.
Deprecated -SuppressCliWarnings in favor of -CliVerbosity NoWarnings. Use of -SuppressCliWarnings will give a
deprecation warning.
Argument completers for -Subscription and -ResourceGroup provide a better description for each completion result.
Documentation updates and fixes.